+++ apps/vesting/src/components/vesting/header/Header.tsx
@@ -6,17 +6,25 @@ import { useState } from "react";
 import { FundVestingAccount } from "./FundVestingAccount";
 import { EnableVestingModal } from "./EnableVestingModal";
 import { useTranslation } from "next-i18next";
+import ApproveFunding from "./ApproveFunding";
 
 export const Header = () => {
   const handleConfirmClick = () => {
     // TODO: open modal for creating vesting account
     setShowModal(true);
-    setModalContent(<FundVestingAccount />);
+    setModalContent(
+      <FundVestingAccount
+        onClose={() => {
+          setShowModal(false);
+        }}
+      />,
+    );
   };
 
   const value = useSelector((state: StoreType) => state.wallet.value);
   const [showModal, setShowModal] = useState(false);
   const [showEnableModal, setShowEnableModal] = useState(false);
+  const [showApproveModal, setShowApproveModal] = useState(false);
 
   const [modalContent, setModalContent] = useState<JSX.Element>(<></>);
   const { t } = useTranslation();
@@ -29,6 +37,14 @@ export const Header = () => {
             text={t("enable.header.button.title")}
             onClick={() => {
               setShowEnableModal(true);
+            }}
+            disabled={!value.active}
+          />
+          <ConfirmButton
+            className="w-fit normal-case"
+            text={t("approve.header.button.title")}
+            onClick={() => {
+              setShowApproveModal(true);
             }}
             disabled={!value.active}
           />
@@ -48,7 +64,24 @@ export const Header = () => {
           setShowEnableModal(false);
         }}
       >
-        <EnableVestingModal />
+        <EnableVestingModal
+          onClose={() => {
+            setShowEnableModal(false);
+          }}
+        />
+      </Modal>
+
+      <Modal
+        show={showApproveModal}
+        onClose={() => {
+          setShowApproveModal(false);
+        }}
+      >
+        <ApproveFunding
+          onClose={() => {
+            setShowApproveModal(false);
+          }}
+        />
       </Modal>
 
       <Modal